pyproject.toml
[tool.poetry]
name = "llm-novelist"
version = "0.1.0"
description = "An AI-powered novel generation tool"
authors = ["Dongyan Zhou <meetzdy@gmail.com>"]
readme = "README.md"
packages = [{include = "llm_novelist", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
ebooklib = "^0.18"
loguru = "^0.7.2"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py38" 